Use react/http as HTTP client in Agent
While cURL does it's job just file, its usage in the Agent will block the event loop for the duration of the push. That means it might miss metrics. Missing a bit isn't that much of an issue, but blocking the event loop every X seconds for the duration of that HTTP call, defeats part of the purpose of putting an agent between the application and the metrics service.